Split multipart
