+++ frictionless/plugins/multipart/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import MultipartPlugin
+from .control import MultipartControl
+from .loader import MultipartLoader
